flask_site/requirements.txt
flask
flask-session
flask-htmx
python-dotenv
WTForms
Flask-WTF
